Test accuracy is 92%!
The reason it was broken before was because weights were initialized as all positive.
This caused the sigmoid output of the first hidden layer to be all activated ~1 which
made every observation look the same. So the neural network just trained the 2nd layer
to take that output and predict the most common digit.